Removed sensitivityEvaluationManager class and moved its capabilities into SensitivityEvaluation
